fix(ci): resolve compilation and clippy errors after rebase
- Fix ambiguous .into() on string literals in asset.rs (primitive_types conflict)
- Fix AccountId version mismatch in price_transformer.rs (near_account_id v1 vs v2)
- Fix unnecessary cast in wad.rs
- Add workspace.package.version for client crate inheritance
- Add missing_errors_doc/missing_panics_doc to workspace lint allows
- Fix ~65 clippy warnings in client/vault (format strings, map_or, casts, dead code, etc.)
- Regenerate Cargo.lock for client crate dependencies

+++ common/src/vault/wad.rs
@@ -257,7 +257,7 @@ impl<'de> Deserialize<'de> for Wad {
         let mut acc = U256::zero();
         let ten = U256::from(10u8);
         for b in s.bytes() {
-            let d = (b - b'0') as u8;
+            let d = b - b'0';
             acc = acc
                 .checked_mul(ten)
                 .ok_or_else(|| near_sdk::serde::de::Error::custom("Wad decimal overflow"))?;
